Move components related to api-adresse in api-adresse
