+++ components/api-doc/api-adresse/technical-doc/path.js
@@ -1,7 +1,7 @@
 import React from 'react'
 import PropTypes from 'prop-types'
 
-import theme from '../../../styles/theme'
+import theme from '../../../../styles/theme'
 
 import ExpandableMenu from '../expandable-menu'
 import ParamsTable from './params-table'
