+++ components/api-doc/api-adresse/try-container.js
@@ -1,7 +1,7 @@
 import React from 'react'
 import PropTypes from 'prop-types'
 
-import theme from '../../styles/theme'
+import theme from '../../../styles/theme'
 
 import NotificationApi from './notification-api'
